PDF veikia
Ir klausimyne ir kitur galima parsisiusti PDF su normaliai isdestyta informacija + rodomi sistemu linkai
