+++ backend/Systems/Update.js
@@ -1,5 +1,6 @@
 export const systemsToUpdate = [
-    { name: 'Service Area 1', link: 'value 1', description: 'updated long description' },
-    { name: 'Service Area 2', link: 'updated value', description: 'updated long description'},
-    { name: 'Nonexistent Service Area', link: 'updated value', description: 'updated long description'}
+    { name: 'LB MANAGEMENT', link: 'https://www.jung-group.com/de-DE/Service/Tools-Services/Schalter-Konfigurator/', description: 'updated long description' },
+    { name: 'eNet SMART HOME', link: '', description: 'updated long description'},
+    { name: 'KNX VISU PRO SERVER', link: 'https://www.jung-group.com/de-DE/Service/Tools-Services/KNX-Konfigurator/', description: 'updated long description'},
+    { name: 'JUNG HOME', link: 'https://www.jung-group.com/de-DE/Service/Tools-Services/KNX-Konfigurator/', description: 'updated long description'}
 ];
